+++ Craftsman/Builders/Projects/IntegrationTestsCsProjBuilder.cs
@@ -46,7 +46,7 @@ public class IntegrationTestsCsProjBuilder
     <PackageReference Include=""NUnit"" Version=""3.13.3"" />
     <PackageReference Include=""NUnit3TestAdapter"" Version=""4.2.1"" />
     <PackageReference Include=""Microsoft.NET.Test.Sdk"" Version=""17.1.0"" />
-    <PackageReference Include=""Respawn"" Version=""5.0.1"" />
+    <PackageReference Include=""Respawn"" Version=""6.0.0"" />
   </ItemGroup>
 
   <ItemGroup>