+++ CHANGELOG.md
@@ -80,7 +80,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Query list uses `AsNoTracking`
 * No more fake read DTO
 * Allow custom schema id's on `AddSwaggerGen`
-* Extended built in validation exception with new overrides to allow for more elegant calls in code
+* Respawn to v6 with simplified `TestFixture` setup
 
 ### Fixed